Do not try to fetch the gender automatically
This is more trouble than it's worth.

On Google, to get access to the gender now you need to ask for permission for user contacts and we don't want that
On Facebook, getting gender now requires a separate permission, access to which requires app review. And automatically filling forms is mentioned under disallowed usages...

Apparently, gender became a top secret thing :P

Closes #104
Related #102, #105